requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
python-dotenv
telethon
requests
speedtest-cli
google-generativeai
pyfiglet
cowsay
